01. Python/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pyserial==3.5
terminal-colors==3.0.0
